fetch.pyをリファクタリングしてmain.pyに変更
