+++ docs/index.md
@@ -1,8 +1,5 @@
-# Open Data 整形済み最新データ
-[テスト](data/20230928-22:07:51-regular.xml) \([source](https://www.data.jma.go.jp/developer/xml/feed/regular.xml)\)
-
-[東京オープンデータカタログ](data/20230928-22:07:54-TokyoOpenDataCatalog) \([source](https://catalog.data.metro.tokyo.lg.jp/api/3/action/package_search?q=*:*&rows=999999&start=0)\)
-
-[東京都認証保育所一覧](data/20230928-22:07:57-ninshouichiran050801.xlsx) \([source](https://www.fukushi.metro.tokyo.lg.jp/kodomo/hoiku/ninsyo/ichiran.files/ninshouichiran050801.xlsx)\)
-
-last update: 20230928-22:07:57
+# Open Data 整形済み最新データ  
+[テスト](data/20230928-23:04:37-regular.xml) \([source](https://www.data.jma.go.jp/developer/xml/feed/regular.xml)\)  
+[東京オープンデータカタログ](data/20230928-23:04:40-TokyoOpenDataCatalog) \([source](https://catalog.data.metro.tokyo.lg.jp/api/3/action/package_search?q=*:*&rows=999999&start=0)\)  
+[東京都認証保育所一覧](data/20230928-23:04:42-ninshouichiran050801.xlsx) \([source](https://www.fukushi.metro.tokyo.lg.jp/kodomo/hoiku/ninsyo/ichiran.files/ninshouichiran050801.xlsx)\)  
+last update: 20230928-23:04:42
